+++ modules/codegen-v2/src/main/kotlin/club/doki7/babel/cdecl/tokenize.kt
@@ -249,6 +249,8 @@ internal val knownTriviaMacros = setOf(
     "SDL_SCANF_FORMAT_STRING",
     "SDL_MALLOC",
 
+    "SHADERC_EXPORT",
+
     "extern"
 )