+++ modules/openal/src/main/java/module-info.java
@@ -13,7 +13,7 @@
 /// <dependency>
 ///     <groupId>club.doki7</groupId>
 ///     <artifactId>openal</artifactId>
-///     <version>0.3.3</version>
+///     <version>0.3.4</version>
 ///     <scope>compile</scope>
 /// </dependency>
 /// ```