+++ modules/vma/src/main/java/module-info.java
@@ -12,7 +12,7 @@
 /// <dependency>
 ///     <groupId>club.doki7</groupId>
 ///     <artifactId>vma</artifactId>
-///     <version>0.3.3</version>
+///     <version>0.3.4</version>
 ///     <scope>compile</scope>
 /// </dependency>
 /// ```